Added Login system and user.cs
basic login system with list of users that can create more users or login, also a way for the program to check if the user is logged in! Also added a user.cs with a basic constructer and a TryLogin that returns a bool, ShowCurrentUser that isnt being used and probably wont!